Use sendRawTransaction RPC endpoint for transfer, claim and contract deployment (#48)
* fix models

* submit transfer via sendrawtransaction

* add show transaction command

* RPC claim

* retrieve neo/gas hashs from neo-express instance

* RPC deploy contract

* calculate neo/gas asset hash locally

* GetAppLog on Show Tx

+++ src/neo-express/Models/UnspentTransaction.cs
@@ -9,7 +9,7 @@ namespace NeoExpress.Models
         public string TransactionId { get; set; } = string.Empty;
 
         [JsonProperty("n")]
-        public long Index { get; set; }
+        public ushort Index { get; set; }
 
         [JsonProperty("value")]
         public decimal Value { get; set; }
